[ignore] small refactor on the toolbox list method (#1992)
